+++ drake/systems/plants/collision/CMakeLists.txt
@@ -6,7 +6,7 @@ if(bullet_FOUND)
   set(drakeCollision_SRC_FILES ${drakeCollision_SRC_FILES} bullet_model.cc)
   set(bullet "yes" CACHE STRING "yes")
 else()
-  set(drakeCollision_SRC_FILES ${drakeCollision_SRC_FILES} UnusableModel.cc)
+  set(drakeCollision_SRC_FILES ${drakeCollision_SRC_FILES} unusable_model.cc)
 endif()
 
 add_library_with_exports(LIB_NAME drakeCollision SOURCE_FILES ${drakeCollision_SRC_FILES})
