+++ drake/systems/plants/collision/unusable_model.h
